requirements.txt
bcolz==1.2.1
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
easydict==1.9
graphviz==0.8.4
idna==2.8
kiwisolver==1.1.0
matplotlib==3.0.3
mxnet==1.4.0.post0
numpy==1.14.6
opencv-contrib-python==4.1.0.25
opencv-python==4.1.0.25
pathlib==1.0.1
Pillow==6.0.0
protobuf==3.7.1
pyparsing==2.4.0
python-dateutil==2.8.0
requests==2.21.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
sklearn==0.0
tensorboardX==1.6
torch==1.0.1.post2
torchvision==0.2.2.post3
tqdm==4.31.1
urllib3==1.24.2
